% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{format_prufer}
\alias{format_prufer}
\title{Format a Prufer sequence list as a string}
\usage{
format_prufer(prufer, type)
}
\arguments{
\item{prufer}{A list of sub-sequences}

\item{type}{The sequence type}
}
\value{
A string with sub-sequences comma-separated, joined by ";"
}
\description{
Format a Prufer sequence list as a string
}
