% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_exports.R
\name{generate_samples_batch}
\alias{generate_samples_batch}
\title{Generate samples in batch using C++ backend}
\usage{
generate_samples_batch(
  Nreps,
  n_sampled,
  nb_mean = 0,
  nb_size = 0,
  components_mean = 0,
  fix_N = 0L,
  fix_k = 0L
)
}
\arguments{
\item{Nreps}{Number of samples}

\item{n_sampled}{Number of sampled nodes}

\item{nb_mean}{Mean of NegBinom proposal for N-n}

\item{nb_size}{Size of NegBinom proposal}

\item{components_mean}{Mean number of components in proposal}

\item{fix_N}{If > 0, fix collapsed forest size to this value}

\item{fix_k}{If > 0, fix number of components to this value}
}
\value{
A list with elements size, components, k1, k2, prufer, type
}
\description{
Generate samples in batch using C++ backend
}
