% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_exports.R
\name{log_likelihood_cpp_wrapper}
\alias{log_likelihood_cpp_wrapper}
\title{Compute log-likelihood using C++ backend}
\usage{
log_likelihood_cpp_wrapper(
  data,
  sequence,
  components,
  seq_type,
  cluster = FALSE
)
}
\arguments{
\item{data}{Integer vector of cluster sizes}

\item{sequence}{List of Prufer sub-sequences}

\item{components}{Number of components}

\item{seq_type}{Sequence type string}

\item{cluster}{Whether to marginalize over subtree structures}
}
\value{
Log-likelihood
}
\description{
Compute log-likelihood using C++ backend
}
