% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{sampling_distribution}
\alias{sampling_distribution}
\title{Create a sampling distribution specification}
\usage{
sampling_distribution(
  nb.mean = NULL,
  nb.size = NULL,
  components.mean = NULL,
  fix.N = NULL,
  fix.k = NULL
)
}
\arguments{
\item{nb.mean}{Mean of negative binomial for forest size (N - n).
Ignored if fix.N is set.}

\item{nb.size}{Size (dispersion) parameter of negative binomial.
Ignored if fix.N is set.}

\item{components.mean}{Mean number of components.
Ignored if fix.k is set.}

\item{fix.N}{If not NULL, fix the collapsed forest size to this value.}

\item{fix.k}{If not NULL, fix the number of components to this value.}
}
\value{
A list with class \code{sampling_dist}
}
\description{
Create a sampling distribution specification
}
